Fix a Simultaneous exception when test_reenter.py is run
The test_reenter.py test starts and stops the event loop several
times. The queue_poller is a singleton only created once and, on
the Mac (and probably on FreeBSD) the self.event_map needs to be
reinitialized. While not seen on Linux the same fix is made for
consistency.